Remove some duplicated code
