+++ cmd/unbounded-storage/src/topology/mod.rs
@@ -5,6 +5,8 @@ mod cores;
 mod filters;
 mod host;
 mod sysfs;
+#[cfg(test)]
+mod testutil;
 
 pub use cores::{
     CorePlan, CorePlanConfig, DiskCpuSlot, NicWorker, NicWorkerGroup, NumaPool, ServingShard,
